vertex-memory-bank/pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true
fallback-version = "0.0.0"

[tool.uv]
cache-keys = [{ file = "pyproject.toml" }, { git = { commit = true, tags = true } }]

[project]
name = "vertex-memory-bank"
dynamic = ["version"]
description = "Vertex AI Memory Bank MCP server"
readme = "README.md"
authors = [
    { name = "zchee", email = "zchee.io@gmail.com" },
]
requires-python = ">=3.14"
dependencies = [
    "google-cloud-aiplatform>=1.73.0",
    "mcp[cli]>=1.0.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
vertex-memory-bank = "vertex_memory_bank:main"

[dependency-groups]
dev = [
    "mypy",
    "pyright",
    "pytest",
    "ruff",
]

[tool.ruff]
src = ["src"]
target-version = "py314"
output-format = "grouped"
respect-gitignore = true
line-length = 120
preview = true
unsafe-fixes = true

format.preview = true
lint.preview = true

# See https://docs.astral.sh/ruff/rules for more information.
lint.select = [
    "ARG", # flake8-unused-arguments
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "D101", # pydocstyle/undocumented-public-class
    "D102", # pydocstyle/undocumented-public-method
    "D103", # pydocstyle/undocumented-public-function
    "E", # pycodestyle (Error)
    "F", # Pyflakes
    "I", # isort
    "INP001", # flake8-no-pep420/implicit-namespace-package
    "N", # pep8-naming
    "SIM", # flake8-simplify
    "TD002", # flake8-todos/missing-todo-author
    "TD006", # flake8-todos/invalid-todo-capitalization
    "TID", # flake8-tidy-imports
    "UP", # pyupgrade
    "W", # pycodestyle (Warning)
]

lint.ignore = [
    "N802", # Function name {name} should be lowercase
    "N805", # First argument of a method should be named self

    # ruff formatter recommends to disable those. see https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191", # Indentation contains tabs
    "E111", # Indentation is not a multiple of
    "D206", # Docstring should be indented with spaces, not tabs
    "E114", # Indentation is not a multiple of {indent_size} (comment)
    "E117", # Over-indented (comment)
    "E501", # Line too long ({width} > {limit})
    "D300", # Use triple double quotes """
    "Q000", # Single quotes found but double quotes preferred
    "Q001", # Single quote multiline found but double quotes preferred
    "Q002", # Single quote docstring found but double quotes preferred
    "Q003", # Change outer quotes to avoid escaping inner quotes
    "COM812", # Trailing comma missing
    "COM819", # Trailing comma prohibited
    "ISC002", # Implicitly concatenated string literals over multiple lines
]

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*" = [
    "D101",
    "D102",
    "D103",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
lines-after-imports = 2
split-on-trailing-comma = true
combine-as-imports = true

[tool.mypy]
python_version = "3.14"
ignore_missing_imports = true
check_untyped_defs = true
sqlite_cache = true
cache_fine_grained = true
strict = true
disable_error_code = [
    "no-untyped-call", # Call to untyped function "xxx" in typed context
    "no-any-return", # Returning Any from function declared to return "xxx"
    "misc", # Class cannot subclass "xxx" (has type "Any")
    "unused-ignore", # Unused "type: ignore" comment
]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
reportUnnecessaryIsInstance = false
reportMissingTypeStubs = false
